+++ src/app/pages/user/desktops/desktops.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../../services/user.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatSnackBar, MatSnackBarDismiss } from '@angular/material/snack-bar';
 import { Desktop, DesktopsResponse, DesktopResponse } from '../../../interfaces/desktop.interface';
 
 @Component({
@@ -28,17 +28,20 @@ export class DesktopsComponent implements OnInit {
   editDesktop(idDesktop: string): void {
 
   }
-  
-  deleteDesktop(idDesktop: string): void {
-    this.userService.deleteDesktop(idDesktop).subscribe((data: DesktopResponse) => {
-      this.snack.open(data.msg, null, { duration: 5000 });
-      this.desktops = this.desktops.filter(desktop => desktop._id != idDesktop);
-    },
-      (err: DesktopResponse) => {
-        this.snack.open(err.msg, null, { duration: 5000 });
-      }
-    )
 
+  deleteDesktop(idDesktop: string): void {
+    this.snack.open('Desea eliminar el escritorio?', 'ELIMINAR', { duration: 10000 }).afterDismissed().subscribe((data: MatSnackBarDismiss) => {
+      if (data.dismissedByAction) {
+        this.userService.deleteDesktop(idDesktop).subscribe((data: DesktopResponse) => {
+          this.snack.open(data.msg, null, { duration: 5000 });
+          this.desktops = this.desktops.filter(desktop => desktop._id != idDesktop);
+        },
+          (err: DesktopResponse) => {
+            this.snack.open(err.msg, null, { duration: 5000 });
+          }
+        )
+      }
+    });
   }
 
   desktopCreated(desktop: Desktop): void {
